pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "google-generativeai>=0.8.6",
    "pytelegrambotapi>=4.29.1",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.40.0",
]
